[FLINK-5497] [tests] Remove duplicated tests for hash tables
This closes #3089
